Project finished
New functionality: keyEvents. When some keys are released (numbers, decimals, operations...) the event is captured to act as if the corresponding button was pressed.

Change ClickEvents on buttons to MouseClickeEvents to fix a bug when the enter key is pressed.

Some methods and fields renamed to conform to the C# standard.

Removed unused imports.